add productRestController

+++ product/src/main/java/com/vnco/fusiontech/product/repository/AttributeRepository.java
@@ -5,6 +5,6 @@ import com.vnco.fusiontech.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface AttributeRepository extends JpaRepository<Attribute, Long> {
+public interface AttributeRepository extends JpaRepository<Attribute, Integer> {
 
 }
